Moves firebase logic to Firebase.js
