+++ frontend/src/Firebase.js
@@ -1,0 +1,68 @@
+import firebase from "firebase/compat/app";
+import * as firebaseui from "firebaseui";
+import "firebaseui/dist/firebaseui.css";
+
+const firebaseConfig = {
+  apiKey: process.env.REACT_APP_FIREBASE_APIKEY,
+  authDomain: "parkcheck-512ae.firebaseapp.com",
+  projectId: "parkcheck-512ae",
+  storageBucket: "parkcheck-512ae.appspot.com",
+  messagingSenderId: "103972727332",
+  appId: "1:103972727332:web:a6aee04b854a56214486b1",
+  measurementId: "G-52F7QV6L6N",
+};
+
+var uiConfig = {
+  callbacks: {
+    signInSuccessWithAuthResult: function (authResult, redirectUrl) {
+      // User successfully signed in.
+      // Return type determines whether we continue the redirect automatically
+      // or whether we leave that to developer to handle.
+      return true;
+    },
+    uiShown: function () {
+      // The widget is rendered.
+      // Hide the loader.
+      document.getElementById("loader").style.display = "none";
+    },
+  },
+  // Will use popup for IDP Providers sign-in flow instead of the default, redirect.
+  signInFlow: "popup",
+  //signInSuccessUrl: "<url-to-redirect-to-on-success>",
+  signInOptions: [
+    // Leave the lines as is for the providers you want to offer your users.
+    //firebase.auth.GoogleAuthProvider.PROVIDER_ID,
+    //firebase.auth.FacebookAuthProvider.PROVIDER_ID,
+    //firebase.auth.TwitterAuthProvider.PROVIDER_ID,
+    //firebase.auth.GithubAuthProvider.PROVIDER_ID,
+    firebase.auth.EmailAuthProvider.PROVIDER_ID,
+    //firebase.auth.PhoneAuthProvider.PROVIDER_ID,
+  ],
+  // Terms of service url.
+  //tosUrl: "<your-tos-url>",
+  // Privacy policy url.
+  //privacyPolicyUrl: "<your-privacy-policy-url>",
+};
+
+export default function Firebase() {
+  const app = firebase.initializeApp(firebaseConfig);
+  var ui = new firebaseui.auth.AuthUI(firebase.auth());
+  ui.start("#firebaseui-auth-container", {
+    signInOptions: [firebase.auth.EmailAuthProvider.PROVIDER_ID],
+    callbacks: {
+      signInSuccessWithAuthResult: function (authResult, redirectUrl) {
+        // User successfully signed in.
+        // Return type determines whether we continue the redirect automatically
+        // or whether we leave that to developer to handle.
+        return true;
+      },
+      uiShown: function () {
+        // The widget is rendered.
+        // Hide the loader.
+        document.getElementById("loader").style.display = "none";
+      },
+    },
+    signInSuccessUrl: "/",
+    // uiConfig,
+  });
+}
