+++ frontend/src/index.js
@@ -8,67 +8,8 @@ import { ThemeProvider } from "styled-components";
 
 import theme from "./assets/theme";
 
-//FB Shit
-import { initializeApp } from "firebase/app";
-import { getAuth } from "firebase/auth";
-
-import firebase from "firebase/compat/app";
-import * as firebaseui from "firebaseui";
-import "firebaseui/dist/firebaseui.css";
-
-const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_APIKEY,
-  authDomain: "parkcheck-512ae.firebaseapp.com",
-  projectId: "parkcheck-512ae",
-  storageBucket: "parkcheck-512ae.appspot.com",
-  messagingSenderId: "103972727332",
-  appId: "1:103972727332:web:a6aee04b854a56214486b1",
-  measurementId: "G-52F7QV6L6N",
-};
-
-const app = firebase.initializeApp(firebaseConfig);
-
-const auth = getAuth(app);
-
-// Initialize the FirebaseUI Widget using Firebase.
-var ui = new firebaseui.auth.AuthUI(firebase.auth());
-
-var uiConfig = {
-  callbacks: {
-    signInSuccessWithAuthResult: function (authResult, redirectUrl) {
-      // User successfully signed in.
-      // Return type determines whether we continue the redirect automatically
-      // or whether we leave that to developer to handle.
-      return true;
-    },
-    uiShown: function () {
-      // The widget is rendered.
-      // Hide the loader.
-      document.getElementById("loader").style.display = "none";
-    },
-  },
-  // Will use popup for IDP Providers sign-in flow instead of the default, redirect.
-  signInFlow: "popup",
-  //signInSuccessUrl: "<url-to-redirect-to-on-success>",
-  signInOptions: [
-    // Leave the lines as is for the providers you want to offer your users.
-    //firebase.auth.GoogleAuthProvider.PROVIDER_ID,
-    //firebase.auth.FacebookAuthProvider.PROVIDER_ID,
-    //firebase.auth.TwitterAuthProvider.PROVIDER_ID,
-    //firebase.auth.GithubAuthProvider.PROVIDER_ID,
-    firebase.auth.EmailAuthProvider.PROVIDER_ID,
-    //firebase.auth.PhoneAuthProvider.PROVIDER_ID,
-  ],
-  // Terms of service url.
-  //tosUrl: "<your-tos-url>",
-  // Privacy policy url.
-  //privacyPolicyUrl: "<your-privacy-policy-url>",
-};
-
-ui.start("#firebaseui-auth-container", {
-  signInOptions: [firebase.auth.EmailAuthProvider.PROVIDER_ID],
-  uiConfig,
-});
+//Firebase logic import
+import Firebase from "./Firebase";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
@@ -85,3 +26,6 @@ root.render(
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
+
+// Call firebase logic
+Firebase();
